Fix search background opacity
The background when the search was opened used to be partially transparent, this restores that and gives it a blur and makes the base background color a dark color for dark mode.

This makes it so you are no longer getting flashed a giant white screen
when you are in dark mode and opened the search.